Admin Container component set up with redux, filter component + DaisyUI next
