+++ src/Private/Pages/HomeAdmin.jsx
@@ -1,9 +1,66 @@
-import React from 'react';
-import { useSelector } from 'react-redux';
+import React, { useEffect } from 'react';
+import { useDispatch, useSelector } from 'react-redux';
 import { AppointmentList } from '../Components/Appointments/AppointmentList';
+import { 
+  setAppointments, 
+  setLoading, 
+  setError,
+  filterAppointments 
+} from '../../store/slices/appointmentSlice';
 
 export const HomeAdmin = () => {
+  const dispatch = useDispatch();
   const { userData } = useSelector(state => state.user);
+  const { 
+    appointments, 
+    filteredAppointments,
+    loading, 
+    error,
+    filters 
+  } = useSelector(state => state.appointments);
+  const { token } = useSelector(state => state.auth);
+
+  useEffect(() => {
+    const fetchAppointments = async () => {
+      dispatch(setLoading(true));
+      try {
+        console.log('🔍 Iniciando fetch de citas...');
+        
+        const response = await fetch('https://psycare-db.onrender.com/admin/appo', {
+          method: 'GET',
+          headers: {
+            'Authorization': `Bearer ${token}`,
+            'Content-Type': 'application/json'
+          }
+        });
+
+        const data = await response.json();
+
+        if (!response.ok) {
+          throw new Error(data.message || 'Error al cargar las citas');
+        }
+
+        const appointments = data.data || [];
+        dispatch(setAppointments(appointments));
+        
+        // Aplicar filtros iniciales si existen
+        if (filters.status !== 'all' || filters.date || filters.type !== 'all') {
+          dispatch(filterAppointments(filters));
+        }
+
+      } catch (err) {
+        console.error('❌ Error:', err.message);
+        dispatch(setError(err.message));
+      }
+    };
+
+    fetchAppointments();
+  }, [dispatch, token, filters]);
+
+  // Determinar qué lista de citas mostrar
+  const appointmentsToShow = filteredAppointments.length > 0 
+    ? filteredAppointments 
+    : appointments;
 
   return (
     <div className="p-8">
@@ -19,7 +76,11 @@ export const HomeAdmin = () => {
 
       {/* Sección de Citas */}
       <div className="bg-white rounded-lg shadow">
-        <AppointmentList />
+        <AppointmentList 
+          appointments={appointmentsToShow}
+          loading={loading}
+          error={error}
+        />
       </div>
     </div>
   );
